requirements.txt
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
pdfminer.six==20221105
pycparser==2.21
